% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiscrimFedorov.R
\name{DiscrimFedWynn}
\alias{DiscrimFedWynn}
\title{Hybrid Algorithm of PSO and L-BFGS for Finding Optimal Discrimination Design}
\usage{
DiscrimFedWynn(
  MODEL_INFO,
  DISTANCE,
  dsLower,
  dsUpper,
  FED_INFO = NULL,
  LBFGS_INFO = NULL,
  seed = NULL,
  verbose = TRUE,
  environment,
  ...
)
}
\description{
Implement the hybridized PSO algorithm to find the optimal discrimination designs
for 2 or more than 2 competing models
}
\references{
Atkinson, A. C. and Fedorov, V. V. (1975a). The design of experiments for discriminating between two rival models. Biometrika, 62(1):57-70.

Lopez-Fidalgo, J., Tommasi, C., and Trandafir, P. C. (2007). An optimal experimental design criterion for discriminating between non-normal models. Journal of the Royal Statistical Society: Series B (Statistical Methodology), 69(2):231-242.
}
