% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiscrimOD.R
\name{emptyModelList}
\alias{emptyModelList}
\title{Create An Empty Model List}
\usage{
emptyModelList(N_model = 2)
}
\arguments{
\item{N_model}{Integer.}
}
\value{
An List.
}
\description{
Please follow the instruction.
}

