% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiscrimOD_assist.r
\name{getAlgInfo}
\alias{getAlgInfo}
\title{Initialize PSO options}
\usage{
getAlgInfo(nSwarm = 64, typePSO = NULL, maxIter = NULL,
  checkConv = NULL, freeRun = NULL, tol = NULL, c1 = NULL, c2 = NULL,
  w0 = NULL, w1 = NULL, w_var = NULL, chi = NULL, vk = NULL,
  typeTopo = NULL, nGroup = NULL, GC_S_ROOF = NULL, GC_F_ROOF = NULL,
  GC_RHO = NULL, Q_cen_type = NULL, Q_a0 = NULL, Q_a1 = NULL,
  Q_a_var = NULL, LcRi_L = NULL, PrSc_P = NULL)
}
\arguments{
\item{nSwarm}{integer. Swarm size.}

\item{typePSO}{integer, PSO type.
\itemize{
\item{0}{Linearly Decreasing Weight PSO (Shi, Y. H.    and Eberhart, R. C., 1998)}
\item{1}{GCPSO (van den Bergh, F. and    Engelbrecht, A. P., 2002)}
\item{2}{Quantum PSO (Sun, J., Feng, B. and Xu, W., 2004)}
\item{3}{LcRiPSO (Bonyadi, M. R., Michalewicz, Z., 2014)}
\item{4}{PrScPSO (Silva, A.,    Neves, A.    and Goncalves, T., 2012)}
}}

\item{maxIter}{integer. Maximal number of iterations.}

\item{checkConv}{logical. The stopping criterion.}

\item{freeRun}{numeric. The persentage of iterations which is free from examing the stopping criterion.}

\item{tol}{numeric. Convergence test. The default is \code{1e-6}.}

\item{c1}{numeric. Default is 2.}

\item{c2}{numeric. Default is 2.}
}
\value{
An List.
}
\description{
Please follow the instruction.
}

