% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiscrimOD_assist.r
\name{getDesignInfo}
\alias{getDesignInfo}
\title{Initialize for information of the optimal design problem}
\usage{
getDesignInfo(D_TYPE = "approx", MODEL_INFO = NULL, dist_func = NULL,
  crit_type = "pair_fixed_true", MaxMinStdVals = NULL, dSupp = 1L,
  nSupp = 2L, dsLower = NULL, dsUpper = NULL)
}
\arguments{
\item{D_TYPE}{string. "exact"}

\item{nSupp}{integer. Number of support points.}
}
\value{
An List.
}
\description{
Please follow the instruction.
}

