% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiscrimOD_assist.r
\name{getFEDInfo}
\alias{getFEDInfo}
\title{Generation function of Fedorov-Wynn algorithm parameter settings}
\usage{
getFEDInfo(FED_MAXIT = 200, FED_TRIM = 5, FED_TRIM_EPS = 1e-04,
  freeRun = 0.5, FED_EPS = 1e-06, FED_ALPHA_GRID = 20)
}
\arguments{
\item{freeRun}{A number between \eqn{[0,1]} that controls the percentage of updating iterations which are free from examining the 
stopping criterion, \eqn{|f'-f|<\varepsilon}
where \eqn{f'} and \eqn{f} are the objective function values in the previous and current iterations, respectively.
The default is 1.00 implying the algorithm will completely ignore the stopping criterion. 
Otherwise, the algorithm checks the stopping criterion after free iterations.}
}
\value{
The list of L-BFGS parameter settings.
}
\description{
Create a list with Fedorov-Wynn algorithm parameters for optimal discrimination design search.
}
\examples{
# Get default settings for Fedorov-Wynn algorithm.
FED_INFO <- getFEDInfo(FED_MAXIT = 200)

}
