% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{T_optimal}
\alias{T_optimal}
\title{Distance measurement function for the T-optiaml criterion}
\usage{
T_optimal(xt, xr)
}
\arguments{
\item{xt}{vector of design points.}

\item{xr}{vector of design weight. Its length must be equal to the length of \code{x} and \code{sum(w)} should be 1.}
}
\value{
Fisher information matrix.
}
\description{
Distance measurement function for the T-optiaml criterion
}
\details{

}
\references{
Dette, H., Kiss, C., Bevanda, M., & Bretz, F. (2010). Optimal designs for the EMAX, log-linear and exponential models. Biometrika, 97(2), 513-518.
}
