% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiscrimOD.R
\name{designCriterion}
\alias{designCriterion}
\title{Calculate Design Criterion Values}
\usage{
designCriterion(DESIGN1, MODEL_INFO, DISTANCE, dsLower, dsUpper,
  MaxMinStdVals = NULL, ALG_INFO = NULL, ...)
}
\arguments{
\item{DESIGN1}{string. "exact"}

\item{ALG_INFO}{list. PSO options.}

\item{DESIGN2}{integer. Number of support points.}

\item{CRIT_TYPE}{string.}

\item{D_INFO}{list. Optimal design problem.}
}
\value{
An List.
}
\description{
Please follow the instruction.
}

