% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiscrimOD.R
\name{designCriterion}
\alias{designCriterion}
\title{Function of criterion value computing}
\usage{
designCriterion(DESIGN1, MODEL_INFO, DISTANCE, dsLower, dsUpper,
  crit_type = "pair_fixed_true", MaxMinStdVals = NULL,
  PSO_INFO = NULL, LBFGS_INFO = NULL, environment, ...)
}
\arguments{
\item{DESIGN1}{matrix. The approximate design.}

\item{MODEL_INFO}{list of information of competing models.
For details, run \code{emptyModelList()} and see the instruction below.}

\item{DISTANCE}{function. The R/C++ function of distance measure.
For T-optimal design, the function is the squared difference of two models means.}

\item{dsLower}{vector. The finite lower bounds of the design space.
Its length should be equal to the dimension of design space.}

\item{dsUpper}{vector. The finite upper bounds of the design space.
Its length should be equal to the dimension of design space.}

\item{crit_type}{string. The name of the case of the discrimination design problem.
The default is 'pair_fixed_true'.}

\item{MaxMinStdVals}{vector. The values of demoninators in the design efficiency calculation for finding max-min discrimination design.}

\item{PSO_INFO}{list. PSO and BFGS options.}
}
\value{
An List.
\itemize{
\item{cri_val}{ the design criterion value of the resulting design in the PSO result or the given design.}
\item{theta2}{ a matrix of the parameters of each input model that result the design criterion value.}
}
}
\description{
Computes the discrimination design criterion value
}
