% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiscrimOD.R
\name{designCriterion}
\alias{designCriterion}
\title{Calculate Design Criterion Values}
\usage{
designCriterion(DESIGN1, MODEL_INFO, DISTANCE, dsLower, dsUpper,
  crit_type = "pair_fixed_true", MaxMinStdVals = NULL, ALG_INFO = NULL,
  environment, ...)
}
\arguments{
\item{DESIGN1}{matrix.}

\item{MODEL_INFO}{list of information of competing models. For details, run \code{emptyModelList()} and see the instruction below.}

\item{DISTANCE}{function.}

\item{dsLower}{vector. The finite lower bounds of the design space. Its length should be equal to the dimension of design space.}

\item{dsUpper}{vector. The finite upper bounds of the design space. Its length should be equal to the dimension of design space.}

\item{crit_type}{string. The name of the case of the discrimination design problem. The default is 'pair_fixed_true'.}

\item{MaxMinStdVals}{vector. The values of demoninators in the design efficiency calculation for finding max-min discrimination design.}

\item{ALG_INFO}{list. PSO and BFGS options.}
}
\value{
An List.
}
\description{
Please follow the instruction.
}

