% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiscrimOD.R
\name{DiscrimOD}
\alias{DiscrimOD}
\title{Hybrid Algorithm of PSO and BFGS for Finding Optimal Discrimination Design}
\usage{
DiscrimOD(MODEL_INFO, DISTANCE, nSupp, dsLower, dsUpper,
  crit_type = "pair_fixed_true", MaxMinStdVals = NULL, ALG_INFO = NULL,
  seed = NULL, verbose = TRUE, environment, ...)
}
\arguments{
\item{MODEL_INFO}{list of information of competing models. For details, run \code{emptyModelList()} and see the instruction below.}

\item{DISTANCE}{function.}

\item{nSupp}{integer. The number fo support points (at least 2).}

\item{dsLower}{vector. The finite lower bounds of the design space. Its length should be equal to the dimension of design space.}

\item{dsUpper}{vector. The finite upper bounds of the design space. Its length should be equal to the dimension of design space.}

\item{crit_type}{string. The name of the case of the discrimination design problem. The default is 'pair_fixed_true'.}

\item{MaxMinStdVals}{vector. The values of demoninators in the design efficiency calculation for finding max-min discrimination design.}

\item{ALG_INFO}{list. PSO and BFGS options.}

\item{seed}{numeric. random seed.  The default is \code{NULL}.}

\item{verbose}{logical. If \code{TRUE}, the PSO will reports the updating progress.}
}
\value{
An List.
}
\description{
Please follow the instruction.
}
\examples{
# Atkinson and Fedorov (1975)
# Two models are given by

# Initialize PSO and BFGS options

# Run Algorithm

}

