% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiscrimOD.R
\name{emptyModelList}
\alias{emptyModelList}
\title{Create An Empty Model List}
\usage{
emptyModelList(N_model = 2)
}
\arguments{
\item{N_model}{Integer. The numebr of competing models including the true model. The default value is 2.}
}
\value{
An List.
\itemize{
\item{True}{ a list of informaiton of true model.}
\itemize{
\item{model}{ the model formula. Input the R function or xPtr function pointer generated by \code{cxxfunction} in the \code{inline} package.}
\item{para}{ the nominal values of parameeters. Input a vector of parameter values.}
}
\item{Rival\code{k}}{ a vector of grid points on the second dimension of the 2-D design space.  For 1_d case, this output can be ignored.}
\itemize{
\item{model}{ the model formula. Input the R function or xPtr function pointer generated by \code{cxxfunction} in the \code{inline} package.}
\item{paraLower}{ the lower bound of teh parameters in the k-th rival model. Input a vector of lower bound values.}
\item{paraUpper}{ the upper bound of teh parameters in the k-th rival model. Input a vector of upper bound values.}
\item{paraInit}{ the initial guess of parameters for the LBFGS algorithm. Input a vector of parameter values.}
}
}
}
\description{
Run \code{emptyModelList()} for instruction of model settings.
}
\examples{
emptyModelList()  # True model and one rival model
emptyModelList(3) # True model and two rival models
emptyModelList(5) # True model and four rival models
}
