% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiscrimOD.R
\name{equivalence}
\alias{equivalence}
\title{Equivalence theorem for discrimination design}
\usage{
equivalence(
  DESIGN = NULL,
  PSO_RESULT = NULL,
  ngrid = 100,
  IFPLOT = FALSE,
  MODEL_INFO,
  DISTANCE,
  dsLower,
  dsUpper,
  crit_type = "pair_fixed_true",
  MaxMinStdVals = NULL,
  PSO_INFO = NULL,
  LBFGS_INFO = NULL,
  ALPHA_PSO_INFO = NULL,
  environment,
  ...
)
}
\arguments{
\item{PSO_RESULT}{list.}

\item{ngrid}{integer. The default is 100.}

\item{IFPLOT}{logical.}

\item{MODEL_INFO}{list of information of competing models. For details, run \code{emptyModelList()} and see the instruction below.}

\item{DISTANCE}{function.}

\item{dsLower}{vector. The finite lower bounds of the design space. Its length should be equal to the dimension of design space.}

\item{dsUpper}{vector. The finite upper bounds of the design space. Its length should be equal to the dimension of design space.}

\item{crit_type}{string. The name of the case of the discrimination design problem. The default is 'pair_fixed_true'.}

\item{MaxMinStdVals}{vector. The values of demoninators in the design efficiency calculation for finding max-min discrimination design.}

\item{PSO_INFO}{list. PSO and BFGS options.}

\item{DESIGN1}{matrix.}
}
\value{
An List.
\itemize{
\item{Grid_1}{ a vector of grid points on the design space.  For 2-D case, this vector is the grid of the first dimension.}
\item{Grid_2}{ a vector of grid points on the second dimension of the 2-D design space.  For 1_d case, this output can be ignored.}
\item{DirDeriv}{ a vector (1-D case) or a matrix (2-D case) of the directional derivative function values on the grid points.}
\item{MAX_DD}{ the maximal value of the directional derivative function in \code{DirDeriv}.}
\item{alpha}{ a vector of weights of efficiency values that satisfies the equivalence theorem for max-min optimal discrimination design.}
}
}
\description{
Computes the values of directional derivative function on an user-defined grid and
returns its maximal value.
}
