% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiscrimOD.R
\name{equivalence}
\alias{equivalence}
\title{Equivalence Theorem}
\usage{
equivalence(DESIGN = NULL, PSO_RESULT = NULL, ngrid = 100,
  IFPLOT = FALSE, MODEL_INFO, DISTANCE, dsLower, dsUpper,
  MaxMinStdVals = NULL, ALG_INFO = NULL, ...)
}
\description{
Equivalence Theorem
}

