% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiscrimOD_assist.r
\name{getLBFGSInfo}
\alias{getLBFGSInfo}
\title{Initialize LBFGS options}
\usage{
getLBFGSInfo(IF_INNER_LBFGS = TRUE, LBFGS_RETRY = 1, LBFGS_MAXIT = 0,
  LBFGS_LM = 6, FVAL_EPS = 0, GRAD_EPS = 1e-05,
  LINESEARCH_MAXTRIAL = 20, LINESEARCH_MAX = 1e+20,
  LINESEARCH_MIN = 1e-20, LINESEARCH_ARMIJO = 1e-04,
  LINESEARCH_WOLFE = 0.9)
}
\arguments{
\item{IF_INNER_LBFGS}{logical.}

\item{LBFGS_RETRY}{integer. The total times of trials in calculating the minimal distance
with randomly generated initial starters. The default is 3.}

\item{LBFGS_MAXIT}{integer. The maximal number of iterations of the L-BFGS algorithm.
The default is 100.}

\item{LBFGS_LM}{integer. The number of corrections to approximate the inverse hessian matrix.
The default is 6.}

\item{FVAL_EPS}{numeric. The value of \eqn{\varepsilon_f} in the stopping criterion,
\eqn{|f'-f|/f<\varepsilon_f} where \eqn{f'} and \eqn{f} are the objective function values
in the previous and current iterations, respectively.
The default is 0 which means not using this stopping criteiron.}

\item{GRAD_EPS}{numeric. The value of \eqn{\varepsilon_g} in the stopping criterion,
\eqn{\|g\|/\|x\|<\varepsilon_g} where \eqn{g} is the gradient vector, \eqn{x} is the current position
and \eqn{\|a\|=\sqrt{a^\top a}}. The default is 1e-5.}

\item{LINESEARCH_MAXTRIAL}{integer The maximal trial of the backtrackign line search. The default is 50.}

\item{LINESEARCH_MAX}{numeric. The initial step size in the line search. The default is 1.}

\item{LINESEARCH_MIN}{numeric. The minimal step size in the line search. The default is 1e-9.}

\item{LINESEARCH_ARMIJO}{numeric. The parameter to
control the accuracy of the backtracking line search algorithm.
The default is \code{1e-4}. The value should be positive and smaller than 0.5.}

\item{LINESEARCH_WOLFE}{numeric. The persentage of the decreasement on the step size in each
iteration of the line search. The default is 0.618.}
}
\value{
The list of algorithm setting parameters as the input of \code{DiscrimOD}.
}
\description{
Please follow the instruction.
}
\examples{
lbfgsInfo <- getLBFGSInfo(LBFGS_RETRY = 2)

}
